requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pandas==2.1.1
numpy==1.26.0
scikit-learn==1.3.1
xgboost==2.0.0
python-multipart==0.0.6
matplotlib==3.8.0
seaborn==0.13.0
joblib==1.3.2
google-generativeai==0.3.1
python-dotenv==1.0.0
pydantic==2.4.2